refactor: move impl for home to features directory
