+++ app/src/features/home/index.ts
@@ -1,0 +1,3 @@
+export * from "./components/faq";
+export * from "./components/features";
+export * from "./components/hero";
